refactor(kubearchive): extract common utils for filter-transform
Extrat common utils for kubearchive filter transform from
"pipeline-run-filter-transform" and "task-run-filter-transforms" files.
